# docs/architecture/contracts/rest/openapi.yaml
openapi: 3.1.0
info:
  title: Mintkey Admin REST API
  version: "0.1.0-experimental"
  summary: Operator-facing control-plane HTTP surface for Mintkey.
  description: |
    Mintkey is a credential broker for AI agents. This API is consumed by:
      - The Admin Console (AdminJS) for sensitive operations.
      - Operator-owned tooling (CI, scripts) using API tokens.
      - Internal subscribers calling `/v1/changes` for reconciliation
        (per ADR-0010 change-channel model).

    Stability tier: **experimental**. Breaking changes will bump the major
    `info.version`; surface stability is per-resource and called out in each
    operation `description` where it differs from the document-level tier.

    ## ID conventions

    All identifiers are **ULIDs** (Crockford base32, 26 characters) prefixed
    by their resource type. Tenants are addressed at the wire by their prefixed ULID `tenant_…` on
    every entity for RLS purposes (ADR-0008).

    | Prefix          | Resource                          |
    | --------------- | --------------------------------- |
    | `tenant_…`      | Tenant                            |
    | `operator_…`    | Operator                          |
    | `agent_…`       | Agent                             |
    | `svc_…`         | Service                           |
    | `cred_…`        | Credential (key-version envelope) |
    | `perm_…`        | PermissionGrant                   |
    | `audit_…`       | AuditEvent                        |
...

...
servers:
...
security:
  - OperatorSession: []
  - OperatorBearer: []
paths:
...
  /v1/tenants/{tenant_id}/services:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      tags: [Services]
      operationId: listServicesByTenant
      summary: List services in a tenant.
      description: Paginated list of registered services scoped to the tenant.
      parameters:
        - $ref: "#/components/parameters/AfterCursor"
        - $ref: "#/components/parameters/Limit"
        - in: query
          name: name_contains
          schema:
            type: string
            maxLength: 200
          description: Substring filter (case-insensitive) on the service name.
        - in: query
          name: q
          required: false
          schema:
            type: string
            maxLength: 200
          description: >
            Case-insensitive substring search across name, slug, description,
            and base_url. LIKE metacharacters (%, _) in the value are treated
            literally.
      responses:
        "200":
          description: A page of services.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicePage"
              examples:
                two-services:
                  value:
                    data:
                      - id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S1"
                        tenant_id: "tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
                        name: "crm"
                        display_name: "Acme CRM"
                        description: "Internal CRM API."
                        base_url: "https://crm.acme.example/api"
                        auth_scheme: "bearer_token"
                        actions: ["read:contacts", "write:contacts"]
                        openapi_url: "https://crm.acme.example/openapi.json"
                        current_key_version: 3
                        status: "active"
                        created_at: "2026-04-25T09:00:00Z"
                        updated_at: "2026-05-08T16:22:00Z"
                      - id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S2"
                        tenant_id: "tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
                        name: "billing"
                        display_name: "Stripe billing"
                        description: "Stripe REST API."
                        base_url: "https://api.stripe.com"
                        auth_scheme: "api_key_header"
                        actions: ["read:invoices", "write:invoices"]
                        openapi_url: null
                        current_key_version: 1
                        status: "active"
                        created_at: "2026-04-26T10:00:00Z"
                        updated_at: "2026-04-26T10:00:00Z"
                    next_cursor: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Services]
      operationId: createService
      summary: Register a new service.
      description: |
        Operator must have at least `Admin` or `AgentOwner` role in the
        tenant. Service `name` must be unique within the tenant. Egress
        allowlisting is anchored to `base_url` per ADR-0007.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
            example:
              name: "crm"
              display_name: "Acme CRM"
              description: "Internal CRM API."
              base_url: "https://crm.acme.example/api"
              auth_scheme: "bearer_token"
              actions: ["read:contacts", "write:contacts"]
              openapi_url: "https://crm.acme.example/openapi.json"
      responses:
        "201":
          description: Service registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
              example:
                id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S1"
                tenant_id: "tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
                name: "crm"
                display_name: "Acme CRM"
                description: "Internal CRM API."
                base_url: "https://crm.acme.example/api"
                auth_scheme: "bearer_token"
                actions: ["read:contacts", "write:contacts"]
                openapi_url: "https://crm.acme.example/openapi.json"
                current_key_version: 0
                status: "active"
                created_at: "2026-05-10T14:00:00Z"
                updated_at: "2026-05-10T14:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/tenants/{tenant_id}/services/{service_id}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/ServiceId"
    get:
      tags: [Services]
      operationId: getService
      summary: Describe a service.
      responses:
        "200":
          description: The service.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
              example:
                id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S1"
                tenant_id: "tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
                name: "crm"
                display_name: "Acme CRM"
                description: "Internal CRM API."
                base_url: "https://crm.acme.example/api"
                auth_scheme: "bearer_token"
                actions: ["read:contacts", "write:contacts"]
                openapi_url: "https://crm.acme.example/openapi.json"
                current_key_version: 3
                status: "active"
                created_at: "2026-04-25T09:00:00Z"
                updated_at: "2026-05-08T16:22:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Services]
      operationId: updateService
      summary: Update non-credential service metadata.
      description: |
        `auth_scheme` cannot be changed after creation; rotate the credential
        and recreate the service if the auth scheme changes upstream. Updates
        to `base_url` are accepted but trigger a Kong-syncer push.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
            example:
              display_name: "Acme CRM (prod)"
              actions: ["read:contacts", "write:contacts", "delete:contacts"]
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
              example:
                id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S1"
                tenant_id: "tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
                name: "crm"
                display_name: "Acme CRM (prod)"
                description: "Internal CRM API."
                base_url: "https://crm.acme.example/api"
                auth_scheme: "bearer_token"
                actions: ["read:contacts", "write:contacts", "delete:contacts"]
                openapi_url: "https://crm.acme.example/openapi.json"
                current_key_version: 3
                status: "active"
                created_at: "2026-04-25T09:00:00Z"
                updated_at: "2026-05-10T15:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Services]
      operationId: deleteService
      summary: Soft-delete a service.
      description: |
        Service is moved to `status = "deleted"`. Active credentials are
        revoked; outstanding tokens fail at the proxy on the next request
        once the change channel propagates.
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  ###########################################################################
  # Credentials
  ###########################################################################
...
  /v1/services:
    get:
      tags: [Services]
      operationId: listServicesImplicit
      summary: (Implicit) List services in the active tenant.
      description: |
        Resolves the active tenant from the operator session cookie or from
        the tenant most-recently selected via `?tenant=…` on the bearer
        token. Equivalent to `GET /v1/tenants/{active_tenant}/services`.
      parameters:
        - $ref: "#/components/parameters/AfterCursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of services.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicePage"
              example:
                data: []
                next_cursor: null
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Services]
      operationId: createServiceImplicit
      summary: (Implicit) Register a service in the active tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
              example:
                id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S5"
                tenant_id: "tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
                name: "billing"
                display_name: "Stripe billing"
                description: ""
                base_url: "https://api.stripe.com"
                auth_scheme: "api_key_header"
                actions: ["read:invoices"]
                openapi_url: null
                current_key_version: 0
                status: "active"
                created_at: "2026-05-10T14:00:00Z"
                updated_at: "2026-05-10T14:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /v1/tenants/{tenant_id}/services/{service_id}/test:
    post:
      operationId: testRunService
      tags: [Services]
      summary: Test the registered service using the current credential.
      description: |
        Operator-initiated synchronous test (F-OP-03). The Admin API fetches
        the current credential from the Vault Adapter, builds a request per
        the service's `auth_scheme`, sends it to `base_url + path`, and
        returns the result. Subject to per-service rate limit (default
        10/min). Emits `service.test_executed` (ADR-0017 C-09).
        Egress allowlisting per ADR-0007 applies (RFC1918 rejected unless
        explicitly allowed for the service).
      security:
        - AdminUiSignedRequest: []
        - OperatorBearer:       []
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/ServiceId"
        - $ref: "#/components/parameters/CsrfHeaderRequired"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TestRunRequest" }
            example:
              method: GET
              path: "/health"
              timeout_ms: 5000
      responses:
        "200":
          description: Test executed; see body for ok/error and timing.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestRunResponse" }
              examples:
                ok:
                  summary: Backend responded 200.
                  value:
                    ok: true
                    status_code: 200
                    latency_ms: 42
                    response_body_truncated: '{"status":"ok"}'
                upstream_unauthorized:
                  summary: Backend rejected the credential.
                  value:
                    ok: false
                    status_code: 401
                    latency_ms: 38
                    response_body_truncated: '{"error":"invalid api key"}'
                timeout:
                  summary: Upstream timed out.
                  value:
                    ok: false
                    latency_ms: 5000
                    error: "timeout"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: |
            Test rejected before sending (forbidden destination, invalid
            input). Body is application/problem+json with mintkey:code
            either `forbidden_destination` or `validation_failed`.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
components:
...
  securitySchemes:
    OperatorSession:
      type: apiKey
      in: cookie
      name: mintkey_session
      description: |
        HttpOnly Secure SameSite=Strict session cookie issued by
        `/v1/auth/internal-login` or `/v1/auth/callback`.
    OperatorBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Operator-owned API token (PAT) for machine-to-machine calls. Issued
        from the Admin Console; not exposed in this OpenAPI document. Sent
        as `Authorization: Bearer …`.
    AgentApiKey:
      type: http
      scheme: bearer
      bearerFormat: mintkey-agent-api-key
      description: |
        Long-lived Agent API Key (`mk_live_…`). Used by the few agent-facing
        REST endpoints; most agent traffic flows through MCP instead.

    AdminUiSignedRequest:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Ed25519 JWT signed by the AdminJS service. Required on every endpoint
        AdminJS calls on behalf of an operator. Carries claims:
          iss: "mintkey/admin-ui"
          sub: <operator_id>
          tnt: <tenant_id>
          aud: "mintkey/admin-api"
          iat, exp (60s TTL)
          jti (UUID; replay-checked against admin_request_jti per ADR-0016.1)
        FastAPI verifies the signature against AdminJS's public key fetched from
        the Vault Adapter under credential type 'admin_ui_signing_key'.
        Per ADR-0014.6 + ADR-0017.1.
    ServiceIdentity:
...